tests: hide tf* tests behind an env var
